+++ CHANGELOG.md
@@ -74,7 +74,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixing infinite loop when retrying in audio and video integ tests
 - Make sure both participants in audio and video tests reach finish state before retrying
 - Trigger videoSendBandwidthDidChange and videoReceiveBandwidthDidChange for Safari
-- Do not disconnect video element with different srcObj when destroying video tile 
+- Do not disconnect video element with different srcObj when destroying video tile
+- Stop content share if the media stream end
 
 ### Removed
 - Remove SDP class withPlanBSimulcast method